requirements.txt
# Mini-Project 1: Image Classification - Caltech-101
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
scikit-learn>=1.2.0
scikit-image>=0.21.0
kaggle>=1.5.0
PyYAML>=6.0
tqdm>=4.65.0
joblib>=1.3.0
matplotlib>=3.5.0
seaborn>=0.12.0
